Move classes into logic package as appropriate
