+++ app/src/main/java/com/kamron/pogoiv/Pokefly.java
@@ -56,6 +56,7 @@ import com.kamron.pogoiv.logic.IVScanResult;
 import com.kamron.pogoiv.logic.PokeInfoCalculator;
 import com.kamron.pogoiv.logic.Pokemon;
 import com.kamron.pogoiv.logic.ScanResult;
+import com.kamron.pogoiv.logic.UpgradeCost;
 import com.kamron.pogoiv.widgets.IVResultsAdapter;
 import com.kamron.pogoiv.widgets.PokemonSpinnerAdapter;
